% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euets.R
\name{co2_euets_files}
\alias{co2_euets_files}
\title{List the EU ETS data file versions bundled with this package}
\usage{
co2_euets_files()
}
\value{
A data frame.
}
\description{
Returns a data frame of the DG CLIMA file vintages supported.
When DG CLIMA publishes a newer file, upgrade the package.
}
\examples{
co2_euets_files()
}
\seealso{
Other EU ETS: 
\code{\link{co2_euets_allocations}()},
\code{\link{co2_euets_emissions}()},
\code{\link{co2_euets_installations}()},
\code{\link{co2_euets_price}()},
\code{\link{co2_euets_surrendered}()}
}
\concept{EU ETS}
