% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euets.R
\name{co2_euets_installations}
\alias{co2_euets_installations}
\title{EU ETS installation registry}
\usage{
co2_euets_installations(country = NULL, refresh = FALSE)
}
\arguments{
\item{country}{Optional two-letter country codes.}

\item{refresh}{Re-download? Default \code{FALSE}.}
}
\value{
A data frame of installations.
}
\description{
Returns the registry of EU ETS installations with metadata:
operator name, activity type, country, permitted since, etc.
}
\examples{
\dontrun{
op <- options(carbondata.cache_dir = tempdir())
inst <- co2_euets_installations(country = "PL")
options(op)
}
}
\seealso{
Other EU ETS: 
\code{\link{co2_euets_allocations}()},
\code{\link{co2_euets_emissions}()},
\code{\link{co2_euets_price}()},
\code{\link{co2_euets_surrendered}()}
}
\concept{EU ETS}
